Add Razor pages and enhance data models
This commit includes the addition of Razor Pages to the service container and updates data models and controllers to use ApplicationUserModel instead of IdentityUser. It also enhances validation attributes in the LoginViewModel and comments out the exclusion of Identity models from migrations for future migration support. Minor logging improvements and code formatting fixes are also included.